requirements.txt
streamlit
spotipy
perlin-noise
numpy
pillow
